# openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Api Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  termsOfService: urn:tos
  title: Api Documentation
  version: "1.0"
servers:
- url: //localhost:8080/
...
paths:
...
  /login:
    post:
      operationId: loginUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginReq'
        description: req
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
          description: OK
        "201":
          content: {}
          description: Created
        "401":
          content: {}
          description: Unauthorized
        "403":
          content: {}
          description: Forbidden
        "404":
          content: {}
          description: Not Found
      summary: login
      tags:
      - Авторизация
      x-codegen-request-body-name: req
      x-contentType: application/json
      x-accepts: '*/*'
  /register:
    post:
      operationId: registerUsingPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegReq'
        description: req
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: object
          description: OK
        "201":
          content: {}
          description: Created
        "401":
          content: {}
          description: Unauthorized
        "403":
          content: {}
          description: Forbidden
        "404":
          content: {}
          description: Not Found
      summary: register
      tags:
      - Авторизация
      x-codegen-request-body-name: req
      x-contentType: application/json
      x-accepts: '*/*'
...
components:
  schemas:
...
    LoginReq:
      example:
        password: password
        username: username
      properties:
        password:
          type: string
        username:
          type: string
      title: LoginReq
      type: object
...
    RegReq:
      example:
        password: password
        role: ADMIN
        username: username
      properties:
        password:
          type: string
        role:
          enum:
          - ADMIN
          - USER
          type: string
        username:
          type: string
      title: RegReq
      type: object
...